rename: 파일 구조 정리 및 미션5 리뷰 내용 반영
