+++ src/assets/styles/mediaquery.css
@@ -14,7 +14,7 @@
     grid-template-columns: repeat(1, 1fr);
   }
 
-  .all-items-list {
+  .all-items-container > .all-items-list {
     grid-template-columns: repeat(2, 1fr);
     gap: 8px;
   }
@@ -37,7 +37,7 @@
     gap: 24px;
   }
 
-  .all-items-list {
+  .all-items-container > .all-items-list {
     grid-template-columns: repeat(3, 1fr);
     gap: 40px;
   }
