+++ src/components/common/Header.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import logo from "../assets/images/logo_pandamarket.png";
+import logo from "../../assets/images/logo_pandamarket.png";
 import "./Header.css";
-import "../styles/mediaquery.css";
+import "../../assets/styles/mediaquery.css";
 
 function Header() {
   return (
