+++ src/components/items/PageNavBar.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Main.css";
+import "../Main.css";
 
 function PageNavBar() {
   return (
