--- src/components/pageNavBar.css
@@ -1,17 +1,0 @@
-.page-nav-bar {
-  display: flex;
-  justify-content: center;
-  gap: 12.5px;
-}
-
-.page-nav-btn {
-  width: 40px;
-  height: 40px;
-  border-radius: 40px;
-  border: 1px solid var(--coolgray-200);
-  background-color: var(--fff);
-  color: var(--coolgray-500);
-  font-weight: 600;
-  font-size: 1rem;
-  padding: 12.5px;
-}
