--- src/styles/mediaquery.css
@@ -1,44 +1,0 @@
-/* mobile */
-@media (max-width: 768px) {
-  /* header */
-  .header {
-    padding: 0 16px;
-  }
-
-  /* main */
-  .main {
-    margin: 94px 16px;
-  }
-
-  .best-items-list {
-    grid-template-columns: repeat(1, 1fr);
-  }
-
-  .all-items-container > .all-items-list {
-    grid-template-columns: repeat(2, 1fr);
-    gap: 8px;
-  }
-}
-
-/* tablet */
-@media (min-width: 768px) and (max-width: 1199px) {
-  /* header */
-  .header {
-    padding: 0 24px;
-  }
-
-  /* main */
-  .main {
-    margin: 94px 24px;
-  }
-
-  .best-items-list {
-    grid-template-columns: repeat(2, 1fr);
-    gap: 24px;
-  }
-
-  .all-items-container > .all-items-list {
-    grid-template-columns: repeat(3, 1fr);
-    gap: 40px;
-  }
-}
